pipeline/dagster/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.18.4
    # via dagster
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.13.2
    # via dagster
anyio==4.12.1
    # via
    #   gql
    #   httpx
    #   starlette
    #   watchfiles
apify-client==2.5.0
    # via zillow-pipeline (pyproject.toml)
apify-shared==2.2.0
    # via apify-client
backoff==2.2.1
    # via gql
cachetools==6.2.6
    # via pyiceberg
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   dagster
    #   dagster-cloud-cli
    #   dagster-webserver
    #   pyiceberg
    #   typer
    #   uvicorn
colorama==0.4.6
    # via apify-client
coloredlogs==14.0
    # via dagster
cryptography==46.0.5
    # via pyjwt
dagster==1.12.14
    # via
    #   zillow-pipeline (pyproject.toml)
    #   dagster-cloud
    #   dagster-graphql
    #   dagster-webserver
dagster-cloud==1.12.14
    # via zillow-pipeline (pyproject.toml)
dagster-cloud-cli==1.12.14
    # via dagster-cloud
dagster-graphql==1.12.14
    # via dagster-webserver
dagster-pipes==1.12.14
    # via dagster
dagster-shared==1.12.14
    # via
    #   dagster
    #   dagster-cloud
    #   dagster-cloud-cli
dagster-webserver==1.12.14
    # via zillow-pipeline (pyproject.toml)
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
docstring-parser==0.17.0
    # via dagster
filelock==3.24.3
    # via dagster
fsspec==2026.2.0
    # via
    #   pyiceberg
    #   universal-pathlib
github3-py==4.0.1
    # via dagster-cloud-cli
gql==4.0.0
    # via dagster-graphql
graphene==3.4.3
    # via dagster-graphql
graphql-core==3.2.7
    # via
    #   gql
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
grpcio==1.78.0
    # via
    #   dagster
    #   grpcio-health-checking
grpcio-health-checking==1.78.0
    # via dagster
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
humanfriendly==10.0
    # via coloredlogs
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
impit==0.11.0
    # via apify-client
jinja2==3.1.6
    # via
    #   dagster
    #   dagster-cloud-cli
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via pyiceberg
more-itertools==10.8.0
    # via apify-client
multidict==6.7.1
    # via yarl
packaging==26.0
    # via
    #   dagster-cloud-cli
    #   dagster-shared
    #   deprecation
pathlib-abc==0.5.2
    # via universal-pathlib
pex==2.59.5
    # via dagster-cloud
platformdirs==4.9.2
    # via dagster-shared
postal==1.1.11
    # via zillow-pipeline (pyproject.toml)
postgrest==2.28.0
    # via supabase
prompt-toolkit==3.0.52
    # via questionary
propcache==0.4.1
    # via yarl
protobuf==6.33.5
    # via
    #   dagster
    #   grpcio-health-checking
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   dagster-shared
    #   postgrest
    #   pyiceberg
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyiceberg==0.11.0
    # via storage3
pyjwt==2.11.0
    # via
    #   github3-py
    #   supabase-auth
pyparsing==3.3.2
    # via pyiceberg
pyroaring==1.0.3
    # via pyiceberg
python-dateutil==2.9.0.post0
    # via
    #   github3-py
    #   graphene
    #   strictyaml
python-dotenv==1.2.1
    # via
    #   zillow-pipeline (pyproject.toml)
    #   dagster
    #   uvicorn
pytz==2025.2
    # via dagster
pyyaml==6.0.3
    # via
    #   dagster-cloud-cli
    #   dagster-shared
    #   uvicorn
questionary==2.1.1
    # via
    #   dagster-cloud
    #   dagster-cloud-cli
realtime==2.28.0
    # via supabase
requests==2.32.5
    # via
    #   dagster
    #   dagster-cloud
    #   dagster-cloud-cli
    #   dagster-graphql
    #   github3-py
    #   gql
    #   pyiceberg
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via gql
rich==14.3.2
    # via
    #   dagster
    #   pyiceberg
    #   typer
setuptools==82.0.0
    # via
    #   dagster
    #   dagster-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   dagster
    #   postal
    #   python-dateutil
sqlalchemy==2.0.46
    # via
    #   alembic
    #   dagster
starlette==0.52.1
    # via
    #   dagster-graphql
    #   dagster-webserver
storage3==2.28.0
    # via supabase
strenum==0.4.15
    # via supabase-functions
strictyaml==1.7.3
    # via pyiceberg
structlog==25.5.0
    # via dagster
supabase==2.28.0
    # via zillow-pipeline (pyproject.toml)
supabase-auth==2.28.0
    # via supabase
supabase-functions==2.28.0
    # via supabase
tabulate==0.9.0
    # via dagster
tenacity==9.1.4
    # via pyiceberg
tomli==2.4.0
    # via dagster
tomlkit==0.14.0
    # via dagster-shared
toposort==1.10
    # via dagster
tqdm==4.67.3
    # via dagster
typer==0.24.0
    # via
    #   dagster-cloud
    #   dagster-cloud-cli
typing-extensions==4.15.0
    # via
    #   alembic
    #   dagster-shared
    #   graphene
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   realtime
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via dagster
universal-pathlib==0.3.9
    # via dagster
uritemplate==4.2.0
    # via github3-py
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via dagster-webserver
uvloop==0.22.1
    # via uvicorn
validators==0.35.0
    # via dagster-cloud-cli
watchdog==6.0.0
    # via dagster
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.6.0
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   realtime
    #   uvicorn
yarl==1.22.0
    # via
    #   gql
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-functions
zstandard==0.25.0
    # via pyiceberg

-e .
